Changed skill_level from int to double in Multiple_choice_Maker

+++ src/main/java/com/example/quizCoach/AI/Multiple_Choice_Maker.java
@@ -34,7 +34,7 @@ public class Multiple_Choice_Maker {
     private ArrayNode Array = objectMapper.createArrayNode();
 
 
-    public Multiple_Choice_Maker(String topic, int skill_level){
+    public Multiple_Choice_Maker(String topic, double skill_level){
 
         /*
         Variables required to connect to the AI
